+++ R/font_info.R
@@ -7,7 +7,9 @@
 #' @inheritParams match_font
 #' @param size The pointsize of the font to use for size related measures
 #' @param res The ppi of the size related mesures
-#' @param path,index path an index of a font file to circumvent lookup based on
+#' @param variation A `font_variation` object or a list of them to control
+#' variable fonts
+#' @param path,index path and index of a font file to circumvent lookup based on
 #' family and style
 #' @param bold `r lifecycle::badge("deprecated")` Use `weight = "bold"` instead
 #'
@@ -61,9 +63,18 @@ font_info <- function(
   res = 72,
   path = NULL,
   index = 0,
+  variation = font_variation(),
   bold = deprecated()
 ) {
-  full_length <- max(length(size), length(res))
+  if (is_font_variation(variation)) variation <- list(variation)
+  full_length <- max(
+    length(size),
+    length(res),
+    length(italic),
+    length(weight),
+    length(width),
+    length(variation)
+  )
   if (is.null(path)) {
     if (lifecycle::is_present(bold)) {
       lifecycle::deprecate_soft("1.2.4", "font_info(bold)", "font_info(weight)")
@@ -71,21 +82,24 @@ font_info <- function(
     }
     full_length <- max(
       length(family),
-      length(italic),
-      length(weight),
-      length(width),
       full_length
     )
+    italic <- rep_len(italic, full_length)
+    weight <- rep_len(weight, full_length)
+    width <- rep_len(width, full_length)
     fonts <- match_fonts(
       family = rep_len(family, full_length),
-      italic = rep_len(italic, full_length),
-      weight = rep_len(weight, full_length),
-      width = rep_len(width, full_length)
+      italic = italic,
+      weight = weight,
+      width = width
     )
     path <- fonts$path
     index <- fonts$index
   } else {
     full_length <- max(length(path), length(index), full_length)
+    italic <- rep_len(italic, full_length)
+    weight <- rep_len(weight, full_length)
+    width <- rep_len(width, full_length)
     if (!all(c(length(path), length(index)) == 1)) {
       path <- rep_len(path, full_length)
       index <- rep_len(index, full_length)
@@ -93,9 +107,22 @@ font_info <- function(
   }
   if (length(size) != 1) size <- rep_len(size, full_length)
   if (length(res) != 1) res <- rep_len(res, full_length)
-  if (!all(file.exists(path)))
+  if (!all(file.exists(path))) {
     stop("path must point to a valid file", call. = FALSE)
-  get_font_info_c(path, as.integer(index), as.numeric(size), as.numeric(res))
+  }
+  variation <- add_standard_to_variations(
+    rep_len(variation, full_length),
+    italic = italic,
+    weight = weight,
+    width = width
+  )
+  get_font_info_c(
+    path,
+    as.integer(index),
+    as.numeric(size),
+    as.numeric(res),
+    variation
+  )
 }
 #' Query glyph-specific information from fonts
 #'
@@ -136,8 +163,10 @@ glyph_info <- function(
   res = 72,
   path = NULL,
   index = 0,
+  variation = font_variation(),
   bold = deprecated()
 ) {
+  if (is_font_variation(variation)) variation <- list(variation)
   n_strings <- length(glyphs)
   glyphs <- strsplit(glyphs, '')
   n_glyphs <- lengths(glyphs)
@@ -151,29 +180,46 @@ glyph_info <- function(
       )
       weight <- ifelse(bold, "bold", "normal")
     }
+    italic <- rep_len(italic, n_strings)
+    weight <- rep_len(weight, n_strings)
+    width <- rep_len(width, n_strings)
     fonts <- match_fonts(
       family = rep_len(family, n_strings),
-      italic = rep_len(italic, n_strings),
-      weight = rep_len(weight, n_strings),
-      width = rep_len(width, n_strings)
+      italic = italic,
+      weight = weight,
+      width = width
     )
     path <- rep(fonts$path, n_glyphs)
     index <- rep(fonts$index, n_glyphs)
+    italic <- rep(italic, n_glyphs)
+    weight <- rep(weight, n_glyphs)
+    width <- rep(width, n_glyphs)
   } else {
     if (!all(c(length(path), length(index)) == 1)) {
       path <- rep(rep_len(path, n_strings), n_glyphs)
       index <- rep(rep_len(index, n_strings), n_glyphs)
     }
+    italic <- rep(rep_len(italic, n_strings), n_glyphs)
+    weight <- rep(rep_len(weight, n_strings), n_glyphs)
+    width <- rep(rep_len(width, n_strings), n_glyphs)
   }
   if (length(size) != 1) size <- rep(rep_len(size, n_strings), n_glyphs)
   if (length(res) != 1) res <- rep(rep_len(res, n_strings), n_glyphs)
-  if (!all(file.exists(path)))
+  if (!all(file.exists(path))) {
     stop("path must point to a valid file", call. = FALSE)
+  }
+  variation <- add_standard_to_variations(
+    rep(rep_len(variation, n_strings), n_glyphs),
+    italic = italic,
+    weight = weight,
+    width = width
+  )
   get_glyph_info_c(
     glyphs,
     path,
     as.integer(index),
     as.numeric(size),
-    as.numeric(res)
+    as.numeric(res),
+    variation
   )
 }