+++ src/FontDescriptor.h
@@ -3,12 +3,15 @@
 
 #include <vector>
 #include <cstring>
+#include <cmath>
 
 #include <ft2build.h>
 #include FT_FREETYPE_H
 #include FT_TRUETYPE_TABLES_H
+#include FT_MULTIPLE_MASTERS_H
 
 #include "utils.h"
+#include "ft_cache.h"
 
 enum FontWeight {
   FontWeightUndefined   = 0,
@@ -67,6 +70,27 @@ inline FontWidth get_font_width(FT_Face face) {
   return (FontWidth) os2_table->usWidthClass;
 }
 
+inline FontWeight fixed_to_weight(int weight) {
+  return (FontWeight) int(weight / FIXED_MOD);
+}
+inline int weight_to_fixed(double weight) {
+  return weight * FIXED_MOD;
+}
+
+inline bool fixed_to_italic(int italic) {
+  return italic / FIXED_MOD > 0.5;
+}
+inline int italic_to_fixed(double italic) {
+  return italic < 0.0 ? 0 : (italic > 1.0 ? FIXED_MOD : italic * FIXED_MOD);
+}
+
+inline FontWidth fixed_to_width(int width) {
+  return width <= 0 ? FontWidthUndefined : (FontWidth) int(std::log2(width / FIXED_MOD) * 4.0 + 5.0);
+}
+inline int width_to_fixed(double width) {
+  return width <= 0.0 ? 0 : std::pow(2.0, (width - 5.0) / 4.0) * FIXED_MOD;
+}
+
 struct FontDescriptor {
 public:
   const char *path;
@@ -78,6 +102,10 @@ public:
   FontWidth width;
   bool italic;
   bool monospace;
+  bool variable;
+  bool var_wght;
+  bool var_wdth;
+  bool var_ital;
 
   FontDescriptor() {
     path = NULL;
@@ -89,6 +117,8 @@ public:
     width = FontWidthUndefined;
     italic = false;
     monospace = false;
+    variable = false;
+    add_var_flags();
   }
 
   // Constructor added by Thomas Lin Pedersen
@@ -102,6 +132,8 @@ public:
     this->width = FontWidthUndefined;
     this->italic = italic;
     this->monospace = false;
+    this->variable = false;
+    add_var_flags();
   }
 
   // Constructor added by Thomas Lin Pedersen
@@ -115,10 +147,12 @@ public:
     this->width = width;
     this->italic = italic;
     this->monospace = false;
+    this->variable = false;
+    add_var_flags();
   }
 
   // Constructor added by Thomas Lin Pedersen
-  FontDescriptor(FT_Face face, const char* path, int index) {
+  FontDescriptor(FT_Face face, const char* path, int index, bool variable = false) {
     this->path = copyString(path);
     this->index = index;
     this->postscriptName = FT_Get_Postscript_Name(face) == NULL ? "" : copyString(FT_Get_Postscript_Name(face));
@@ -128,10 +162,12 @@ public:
     this->width = get_font_width(face);
     this->italic = face->style_flags & FT_STYLE_FLAG_ITALIC;
     this->monospace = FT_IS_FIXED_WIDTH(face);
+    this->variable = variable;
+    add_var_flags();
   }
 
   FontDescriptor(const char *path, const char *postscriptName, const char *family, const char *style,
-                 FontWeight weight, FontWidth width, bool italic, bool monospace) {
+                 FontWeight weight, FontWidth width, bool italic, bool monospace, bool variable = false) {
     this->path = copyString(path);
     this->index = 0;
     this->postscriptName = copyString(postscriptName);
@@ -141,10 +177,12 @@ public:
     this->width = width;
     this->italic = italic;
     this->monospace = monospace;
+    this->variable = variable;
+    add_var_flags();
   }
 
   FontDescriptor(const char *path, int index, const char *postscriptName, const char *family, const char *style,
-                 FontWeight weight, FontWidth width, bool italic, bool monospace) {
+                 FontWeight weight, FontWidth width, bool italic, bool monospace, bool variable = false) {
     this->path = copyString(path);
     this->index = index;
     this->postscriptName = copyString(postscriptName);
@@ -154,6 +192,8 @@ public:
     this->width = width;
     this->italic = italic;
     this->monospace = monospace;
+    this->variable = variable;
+    add_var_flags();
   }
 
   FontDescriptor(FontDescriptor *desc) {
@@ -166,6 +206,10 @@ public:
     width = desc->width;
     italic = desc->italic;
     monospace = desc->monospace;
+    variable = desc->variable;
+    var_wght = desc->var_wght;
+    var_wdth = desc->var_wdth;
+    var_ital = desc->var_ital;
   }
 
   const char* get_path() {
@@ -246,13 +290,13 @@ public:
     if (style && !strcmp_no_case(style, other.style))
       return false;
 
-    if (weight && weight != other.weight)
+    if (weight && !var_wght && !other.var_wght && weight != other.weight)
       return false;
 
-    if (width && width != other.width)
+    if (width &&  !var_wdth && !other.var_wdth && width != other.width)
       return false;
 
-    if (italic != other.italic)
+    if (!var_ital && !other.var_ital && italic != other.italic)
       return false;
 
     return true;
@@ -271,6 +315,18 @@ private:
     char *str = new char[strlen(input) + 1];
     strcpy(str, input);
     return str;
+  }
+  void add_var_flags() {
+    if (variable && path != NULL) {
+      FreetypeCache& cache = get_font_cache();
+      if (cache.load_font(path, 0)) {
+        cache.has_axes(var_wght, var_wdth, var_ital);
+        return;
+      }
+    }
+    var_wght = false;
+    var_wdth = false;
+    var_ital = false;
   }
 };