+++ src/mac/FontManagerMac.mm
@@ -40,6 +40,11 @@ static int convertWidth(float unit) {
 void addFontIndex(FontDescriptor* font) { @autoreleasepool {
   static std::map<std::string, int> font_index;
 
+  if (font->variable) {
+    font->index = 0;
+    return;
+  }
+
   std::string font_name(font->postscriptName);
   int font_no;
   std::map<std::string, int>::iterator it = font_index.find(font_name);
@@ -89,6 +94,8 @@ FontDescriptor *createFontDescriptor(CTFontDescriptorRef descriptor) { @autorele
   NSNumber *symbolicTraitsVal = traits[(id)kCTFontSymbolicTrait];
   unsigned int symbolicTraits = [symbolicTraitsVal unsignedIntValue];
 
+  CFArrayRef variations = CTFontCopyVariationAxes(CTFontCreateWithFontDescriptor(descriptor, 0.0, NULL));
+
   FontDescriptor *res = new FontDescriptor(
     [[url path] UTF8String],
     [psName UTF8String],
@@ -97,7 +104,8 @@ FontDescriptor *createFontDescriptor(CTFontDescriptorRef descriptor) { @autorele
     weight,
     width,
     (symbolicTraits & kCTFontItalicTrait) != 0,
-    (symbolicTraits & kCTFontMonoSpaceTrait) != 0
+    (symbolicTraits & kCTFontMonoSpaceTrait) != 0,
+    variations != nullptr && CFArrayGetCount(variations) != 0
   );
   addFontIndex(res);
   return res;
@@ -183,15 +191,36 @@ int metricForMatch(CTFontDescriptorRef match, FontDescriptor *desc) { @autorelea
 
   bool italic = ([dict[(id)kCTFontSymbolicTrait] unsignedIntValue] & kCTFontItalicTrait);
 
+  // See if font has variations for certain traits. If so, ignore these in the metric
+  bool has_var_weight = false;
+  bool has_var_width = false;
+  bool has_var_italic = false;
+  CFArrayRef variations = CTFontCopyVariationAxes(CTFontCreateWithFontDescriptor(match, 0.0, NULL));
+  if (variations != nullptr) {
+    long weight_tag = 2003265652;
+    long width_tag = 2003072104;
+    long italic_tag = 1769234796;
+    long tag_val = 0;
+    for (CFIndex i = 0; i < CFArrayGetCount(variations); ++i) {
+      CFDictionaryRef axis = (CFDictionaryRef)CFArrayGetValueAtIndex(variations, i);
+      CFNumberRef tag = (CFNumberRef) CFDictionaryGetValue(axis, kCTFontVariationAxisIdentifierKey);
+      CFNumberGetValue(tag, kCFNumberLongType, &tag_val);
+      if (tag_val == weight_tag) has_var_weight = true;
+      else if (tag_val == width_tag) has_var_width = true;
+      else if (tag_val == italic_tag) has_var_italic = true;
+    }
+  }
+
   // normalize everything to base-900
   int metric = 0;
-  if (desc->weight)
+  if (!has_var_weight && desc->weight)
     metric += sqr(convertWeight([dict[(id)kCTFontWeightTrait] floatValue]) - desc->weight);
 
-  if (desc->width)
+  if (!has_var_width && desc->width)
     metric += sqr((convertWidth([dict[(id)kCTFontWidthTrait] floatValue]) - desc->width) * 100);
 
-  metric += sqr((italic != desc->italic) * 900);
+  if (!has_var_italic)
+    metric += sqr((italic != desc->italic) * 900);
 
   return metric;
 }}