+++ src/unix/FontManagerLinux.cpp
@@ -104,6 +104,7 @@ FontWidth convertWidth(int width) {
 FontDescriptor *createFontDescriptor(FcPattern *pattern) {
   FcChar8 *path = NULL, *psName = NULL, *family = NULL, *style = NULL;
   int index = 0, weight = 0, width = 0, slant = 0, spacing = 0;
+  FcBool variable = FcFalse;
 
   FcPatternGetString(pattern, FC_FILE, 0, &path);
 
@@ -120,6 +121,8 @@ FontDescriptor *createFontDescriptor(FcPattern *pattern) {
   FcPatternGetInteger(pattern, FC_WIDTH, 0, &width);
   FcPatternGetInteger(pattern, FC_SLANT, 0, &slant);
   FcPatternGetInteger(pattern, FC_SPACING, 0, &spacing);
+  FcPatternGetBool(pattern, FC_VARIABLE, 0, &variable);
+  if (variable == FcTrue) index = 0;
 
   return new FontDescriptor(
     (char *) path,
@@ -130,7 +133,8 @@ FontDescriptor *createFontDescriptor(FcPattern *pattern) {
     convertWeight(weight),
     convertWidth(width),
     slant == FC_SLANT_ITALIC,
-    spacing == FC_MONO
+    spacing == FC_MONO,
+    variable == FcTrue
   );
 }
 
@@ -155,9 +159,9 @@ ResultSet *getAvailableFonts() {
 
   FcPattern *pattern = FcPatternCreate();
 #ifdef FC_POSTSCRIPT_NAME
-  FcObjectSet *os = FcObjectSetBuild(FC_FILE, FC_POSTSCRIPT_NAME, FC_FAMILY, FC_STYLE, FC_WEIGHT, FC_WIDTH, FC_SLANT, FC_SPACING, NULL);
+FcObjectSet *os = FcObjectSetBuild(FC_FILE, FC_POSTSCRIPT_NAME, FC_FAMILY, FC_STYLE, FC_WEIGHT, FC_WIDTH, FC_SLANT, FC_SPACING, FC_VARIABLE, NULL);
 #else
-  FcObjectSet *os = FcObjectSetBuild(FC_FILE, FC_FAMILY, FC_STYLE, FC_WEIGHT, FC_WIDTH, FC_SLANT, FC_SPACING, NULL);
+  FcObjectSet *os = FcObjectSetBuild(FC_FILE, FC_FAMILY, FC_STYLE, FC_WEIGHT, FC_WIDTH, FC_SLANT, FC_SPACING, FC_VARIABLE, NULL);
 #endif
   FcFontSet *fs = FcFontList(NULL, pattern, os);
   ResultSet *res = getResultSet(fs);
@@ -204,9 +208,9 @@ ResultSet *findFonts(FontDescriptor *desc) {
   FcPattern *pattern = createPattern(desc);
 
 #ifdef FC_POSTSCRIPT_NAME
-  FcObjectSet *os = FcObjectSetBuild(FC_FILE, FC_POSTSCRIPT_NAME, FC_FAMILY, FC_STYLE, FC_WEIGHT, FC_WIDTH, FC_SLANT, FC_SPACING, NULL);
+  FcObjectSet *os = FcObjectSetBuild(FC_FILE, FC_POSTSCRIPT_NAME, FC_FAMILY, FC_STYLE, FC_WEIGHT, FC_WIDTH, FC_SLANT, FC_SPACING, FC_VARIABLE, NULL);
 #else
-  FcObjectSet *os = FcObjectSetBuild(FC_FILE, FC_FAMILY, FC_STYLE, FC_WEIGHT, FC_WIDTH, FC_SLANT, FC_SPACING, NULL);
+  FcObjectSet *os = FcObjectSetBuild(FC_FILE, FC_FAMILY, FC_STYLE, FC_WEIGHT, FC_WIDTH, FC_SLANT, FC_SPACING, FC_VARIABLE, NULL);
 #endif
 
   FcFontSet *fs = FcFontList(NULL, pattern, os);
@@ -230,7 +234,7 @@ FontDescriptor *findFont(FontDescriptor *desc) {
 
   FcPatternDestroy(pattern);
   FcPatternDestroy(font);
-  
+
   // No match try using family as postscriptName
   if (res == NULL) {
     desc->postscriptName = desc->family;
@@ -238,10 +242,10 @@ FontDescriptor *findFont(FontDescriptor *desc) {
     pattern = createPattern(desc);
     FcConfigSubstitute(NULL, pattern, FcMatchPattern);
     FcDefaultSubstitute(pattern);
-    
+
     font = FcFontMatch(NULL, pattern, &result);
     res = font ? createFontDescriptor(font) : NULL;
-    
+
     FcPatternDestroy(pattern);
     FcPatternDestroy(font);
   }