+++ NEWS.md
@@ -11,6 +11,9 @@
 * Added `plot_glyph_stats()` to provide visual explanation for various glyph
   measures
 * `font_info()` now returns the PostScript name of the font in the `name` column
+* Added support for variable fonts throughout the package. Other packages will
+  need to be upgraded to take advantage of this. A new function
+  `font_variation()` can be used to define coords for the variation axes
 * Fixed a bug in webfont download on Windows where fontfiles would become
   corrupted (#134)